vuexのmoduleからgettersを分離
