+++ src/store/modules/MapViewModule.ts
@@ -1,219 +1,12 @@
 import { Getters, Mutations, Actions, Module } from 'vuex-smart-module';
-import MapViewState from './MapViewState';
+import { MapViewState } from './MapViewState';
+import { MapViewGetters } from './MapViewGetters';
 import { Map, Spot, SpotInfo, SpotForMap, Bounds, DisplayLevelType, Coordinate, Node } from '@/store/types';
 import { NoDetailMapsError } from '../errors/NoDetailMapsError';
 import { NoDetailMapIdInSpotError } from '../errors/NoDetailMapIdInSpotError';
 import { MapNotFoundError } from '../errors/MapNotFoundError';
 import { SpotNotFoundError } from '../errors/SpotNotFoundError';
 import { mapViewGetters } from '@/store';
-
-
-export class MapViewGetters extends Getters<MapViewState> {
-    get maps(): Map[] {
-        return this.state.maps;
-    }
-
-    get rootMapId(): number {
-        return this.state.rootMapId;
-    }
-
-    get spotInfoIsVisible(): boolean {
-        return this.state.spotInfoIsVisible;
-    }
-
-    get focusedSpot(): {mapId: number, spotId: number} {
-        return this.state.focusedSpot;
-    }
-
-    /**
-     * Mapコンポーネントが扱うマップの範囲を返す
-     * @return マップの範囲
-     */
-    get rootMapBounds(): Bounds {
-        const rootMapIndex: number = this.state.maps.findIndex((m: Map) => m.id === this.state.rootMapId);
-        return this.state.maps[rootMapIndex].bounds;
-    }
-
-    /**
-     * Mapコンポーネントがアイコン表示のために必要なスポットの情報を返す
-     * - vuex-module-decoratorsにおいて引数付きgetterはこの書き方になる
-     * @param  mapId 必要なスポットが存在するマップのID
-     * @return Mapコンポーネントが必要なスポットの情報
-     */
-    public getSpotsForMap(mapId: number): SpotForMap[] {
-        const mapIndex: number = this.state.maps.findIndex((m: Map) => m.id === mapId);
-        const spots: Spot[] = this.state.maps[mapIndex].spots;
-        const spotsForMap: SpotForMap[] = [];
-        spots.forEach((spot) => {
-            spotsForMap.push({
-                id: spot.id,
-                name: spot.name,
-                coordinate: spot.coordinate,
-                shape: spot.shape,
-            });
-        });
-        return spotsForMap;
-    }
-
-    /**
-     * SpotInfoコンポーネントが表示する情報を返す
-     * @return SpotInfoコンポーネントに必要な情報
-     */
-    get infoOfFocusedSpot(): SpotInfo {
-        const parentMapId: number = this.state.focusedSpot.mapId;
-        const spotId: number = this.state.focusedSpot.spotId;
-        const spot: Spot = this.getters.getSpotById({parentMapId, spotId});
-        const spotInfo: SpotInfo = {
-            name:  spot.name,
-        };
-        return spotInfo;
-    }
-
-    /**
-     * マップ配列から,マップIdとスポットIdで指定されたスポットを取得する
-     * @param maps MapViewModuleのマップ配列
-     * @param targetSpot マップIdとスポットIdのオブジェクト
-     * @throw MapNotFoundError 指定されたマップが見つからない場合に発生
-     * @throw SpotNotFoundError 指定されたスポットが見つからない場合に発生
-     */
-    public getSpotById(targetSpot: {parentMapId: number, spotId: number}): Spot {
-        const map: Map | undefined = this.state.maps.find((m: Map) => m.id === targetSpot.parentMapId);
-        if (map === undefined) {
-            throw new MapNotFoundError('Map does not found...');
-        }
-        const spot: Spot | undefined = map.spots.find((s: Spot) => s.id === targetSpot.spotId);
-        if (spot === undefined) {
-            throw new SpotNotFoundError('Spot does not found...');
-        }
-        return spot;
-    }
-
-    /**
-     * 指定されたスポットが詳細マップを持つかどうかを判定する．
-     * @param targetSpot マップのIdとスポットのId
-     * @return スポットが詳細マップを持つならばtrue, 持たないならばfalse
-     */
-    public spotHasDetailMaps(targetSpot: {
-                parentMapId: number,
-                spotId: number,
-    }): boolean {
-        const spot = this.getters.getSpotById(targetSpot);
-        if (spot.detailMapIds.length > 0) {
-            return true;
-        } else {
-            return false;
-        }
-    }
-
-    /**
-     * スポットのもつ詳細マップのうち、最後に参照された詳細マップのIdを返す
-     * @param parentSpot どのマップのどのスポットかを示す情報.
-     * @return lastViewdDetailMapId スポットが持つ詳細マップのうち、最後に参照された詳細マップのId．
-     * まだ参照されていない場合はnullを返す．
-     * @throw NoDetailMapsError スポットが詳細マップを持っていない場合に発生.
-     */
-    public getLastViewedDetailMapId(parentSpot: {parentMapId: number, spotId: number}): number | null {
-        if (this.getters.spotHasDetailMaps(parentSpot) === false) {
-            throw new NoDetailMapsError('This spot has no detail maps...');
-        }
-        const spot = this.getters.getSpotById(parentSpot);
-        const lastViewedDetailMapId: number | null = spot.lastViewedDetailMapId;
-        return lastViewedDetailMapId;
-    }
-
-    /**
-     * ズームレベルによって変化する表示レベルを返す
-     * @return 表示レベル('default' or 'detail')
-     */
-    get displayLevel(): DisplayLevelType {
-        return this.state.displayLevel;
-    }
-
-    /**
-     * - 画面上で表示されている
-     * - 半径〇〇内で最も画面中央に近い
-     * - 詳細マップを持っている
-     * スポットのIDを返す
-     * 条件に当てはまるスポットがない状態である場合nullを返す
-     * @return スポットIDかnull
-     */
-    get idOfCenterSpotInRootMap(): number | null {
-        return this.state.idOfCenterSpotInRootMap;
-    }
-
-     /* 経由するノードidの配列を入力することで経路となるノードの配列を取得
-     * @param nodeIds: 経路となるノードidの配列
-     * @return nodesForNavigation: 経路となるノードの配列
-     */
-    public getNodesForNavigation(nodeIds: number[]): Coordinate[][] {
-        // getterの中身は経路探索に依存しているため、現状テスト用のものを使用
-        // ノードidの配列を入力として必要なノードを検索、配列として返すメソッドが必要
-        const testRoutes: Node[][] = [[
-            {
-                id: 0,
-                mapId: 0,
-                spotId: 0,
-                coordinate: {
-                    lat: 33.595502,
-                    lng: 130.218238,
-                },
-            },
-            {
-                id: 1,
-                mapId: 0,
-                spotId: 1,
-                coordinate: {
-                lat: 33.596502,
-                lng: 130.218238,
-                },
-            },
-            {
-                id: 2,
-                mapId: 0,
-                spotId: 2,
-                coordinate: {
-                lat: 33.596502,
-                lng: 130.219238,
-                },
-            },
-        ],
-        [
-            {
-                id: 0,
-                mapId: 0,
-                spotId: 0,
-                coordinate: {
-                    lat: 33.595502,
-                    lng: 130.218238,
-                },
-            },
-            {
-                id: 2,
-                mapId: 0,
-                spotId: 2,
-                coordinate: {
-                lat: 33.596502,
-                lng: 130.219238,
-                },
-            },
-            {
-                id: 1,
-                mapId: 0,
-                spotId: 1,
-                coordinate: {
-                lat: 33.596502,
-                lng: 130.218238,
-                },
-            },
-        ]];
-        const nodesForNavigation: Coordinate[][] = [];
-        testRoutes.forEach((route: Node[]) => {
-            const wayPoints: Coordinate[] = route.map((wayPoint: Node) => wayPoint.coordinate);
-            nodesForNavigation.push(wayPoints);
-        });
-        return nodesForNavigation;
-    }
-}
 
 export class MapViewMutations extends Mutations<MapViewState> {
     /**
